refactor: harden ESP state management and input timing utilities
Tighten concurrency safety, input validation, and timing precision across ESP and QOL modules.

## Chest ESP
~ Replace single `waitingForChest` counter with `pendingChestSpawns` `ConcurrentLinkedQueue` for multi-event tracking
~ Add `pruneExpiredPendingSpawns` to evict timed-out spawn entries exceeding `CHEST_SPAWN_TIMEOUT_MS`
~ Gate particle scanning on `activeLockpickChests` presence to reduce false positives
~ Update `reset` and area-gate reset to clear `pendingChestSpawns` queue
(`src/main/kotlin/com/hypcro/mining/ChestESP.kt`)

## White Gift ESP
~ Introduce `DISCOVERY_INTERVAL_MS` constant to decouple scan cadence from `EspHelper`
~ Apply local interval in `discoverGifts` for consistent gift discovery timing
(`src/main/kotlin/com/hypcro/jerry/WhiteGiftESP.kt`)

## Party API
~ Return defensive copy in `getPartyMembers` to prevent external mutation
(`src/main/kotlin/com/hypcro/party/PartyApi.kt`)

## Player ESP
~ Switch `parseRgb` to `Long`-based bit shifting to avoid integer overflow on high hex values
(`src/main/kotlin/com/hypcro/player/PlayerESP.kt`)

## Faster RClick Helper
~ Remove mutable `nextIntervalMs` field in favor of computed `currentIntervalMs` per tick
~ Reset `holdStartTime` on missing player to prevent stale input state
(`src/main/kotlin/com/hypcro/qol/FasterRClickHelper.kt`)

+++ src/main/kotlin/com/hypcro/player/PlayerESP.kt
@@ -91,10 +91,10 @@ object PlayerESP {
     private fun parseRgb(hex: String): Triple<Int, Int, Int> {
         val clean = hex.removePrefix("#").trim()
         return try {
-            val num = clean.toInt(16)
-            val r = (num shr 16) and 0xFF
-            val g = (num shr 8) and 0xFF
-            val b = num and 0xFF
+            val num = clean.toLong(16)
+            val r = ((num shr 16) and 0xFF).toInt()
+            val g = ((num shr 8) and 0xFF).toInt()
+            val b = (num and 0xFF).toInt()
             Triple(r, g, b)
         } catch (_: Exception) {
             Triple(0, 255, 255)
